Move .jsonSchemaToReqlFilter() to lib/core
