+++ lib/core/backend.js
@@ -21,7 +21,7 @@ const debug = require('debug')('jellyfish:backend')
 const errors = require('./errors')
 const jsonSchema = require('./json-schema')
 const Cache = require('./cache')
-const utils = require('../utils')
+const utils = require('./utils')
 
 const isInsertionError = (results) => {
 	return results.errors !== 0
