+++ lib/core/utils.js
@@ -61,3 +61,61 @@ exports.mergeStreams = async (keys, createFunction) => {
 
 	return emitter
 }
+
+/**
+ * @summary Rudimentary mapping from json schema to a reql filter object
+ * @function
+ * @public
+ *
+ * @description Maps `const` values from a json schema into reql filter object
+ * @see https://rethinkdb.com/api/javascript/filter/
+ *
+ * @param {Object} schema - json schema
+ * @returns {Object} reql filter object
+ *
+ * @example
+ * const schema = {
+ * 	type: 'object',
+ * 	properties: {
+ * 		type: {
+ * 			const: 'message'
+ * 		},
+ * 		data: {
+ * 			type: 'object',
+ * 			properties: {
+ * 				target: {
+ * 					const: 'foobarbaz'
+ * 				}
+ * 			},
+ * 			required: [ 'target' ],
+ * 			additionalProperties: true
+ * 		},
+ * 	},
+ * 	required: [ 'type', 'data' ],
+ * 	additionalProperties: true
+ * }
+ * const filter = jsonSchemaToReqlFilter(schema)
+ * console.log(filter) // -> {
+ *	type: 'message',
+ *	data: {
+ *		target: 'foobarbaz'
+ *	}
+ * }
+ */
+exports.jsonSchemaToReqlFilter = (schema) => {
+	return _.reduce(
+		schema.properties,
+		(result, value, key) => {
+			if (value.const) {
+				result[key] = value.const
+			}
+
+			if (value.type === 'object') {
+				result[key] = exports.jsonSchemaToReqlFilter(value)
+			}
+
+			return result
+		},
+		{}
+	)
+}
